# the-corporations-email/openapi.yaml
openapi: 3.0.3
info:
  title: The Corporation's Email API
  description: Internal email system for "The Corporations" - a REST API for sending, retrieving, and managing emails with support for threaded conversations.
  version: 1.0.0
  contact:
    name: The Corporations
servers:
  - url: http://localhost:60061
    description: Local development server
...
paths:
...
  /mail:
    get:
      tags:
        - Mail
      summary: Get inbox
      description: |
        Get all email subjects for a user's inbox. Only shows emails where the viewer
        is a recipient or sender. Excludes emails deleted by the viewer.

        **Pagination:** 10 items per page

        **Sorting:** By timestamp descending (most recent first)
      operationId: getInbox
      parameters:
        - $ref: '#/components/parameters/viewer'
        - $ref: '#/components/parameters/page'
      responses:
        '200':
          description: Inbox retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EmailSummary'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                missingViewer:
                  summary: Missing viewer parameter
                  value:
                    error: "Missing required 'viewer' query parameter"
                    code: MISSING_VIEWER
                invalidViewer:
                  summary: Empty viewer parameter
                  value:
                    error: "Viewer parameter cannot be empty or whitespace"
                    code: INVALID_VIEWER
                invalidPage:
                  summary: Invalid page number
                  value:
                    error: "Invalid page number: '0'. Must be a positive integer."
                    code: INVALID_PAGE

    post:
      tags:
        - Mail
      summary: Send email
      description: |
        Send a new email. All names are normalized to lowercase.

        **Auto-generated fields:**
        - `id`: New UUID
        - `timestamp`: Current datetime (ISO 8601 UTC)
        - `readBy`: Empty array
        - `deletedBy`: Empty array

        **Reply behavior:** If `isResponseTo` is provided and subject doesn't start with "Re: ", it's automatically prepended.
      operationId: sendEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendEmailRequest'
      responses:
        '201':
          description: Email sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    description: ID of the created email
                  message:
                    type: string
                    example: Email sent successfully
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '415':
          description: Unsupported media type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "Unsupported media type: 'text/plain'. Expected 'application/json; charset=utf-8'"
                code: UNSUPPORTED_MEDIA_TYPE
  /mail/{mail_id}:
    get:
      tags:
        - Mail
      summary: Get email detail
      description: |
        Get a specific email with full content, plus summaries of other emails in the conversation thread.

        **Side effects:** Automatically marks the email as read for the viewer.

        **Thread pagination:** 20 items per page

        **Thread includes ALL emails** regardless of delete status.
      operationId: getEmail
      parameters:
        - name: mail_id
          in: path
          required: true
          description: UUID of the email to retrieve
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/viewer'
        - name: thread_page
          in: query
          description: Page number for thread summaries (1-indexed)
          schema:
            type: integer
            minimum: 1
            default: 1
      responses:
        '200':
          description: Email retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    $ref: '#/components/schemas/EmailDetail'
                  thread:
                    type: array
                    description: Other emails in the thread (excludes requested email)
                    items:
                      $ref: '#/components/schemas/ThreadSummary'
                  thread_pagination:
                    $ref: '#/components/schemas/ThreadPagination'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Email not found or deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                notFound:
                  summary: Email does not exist
                  value:
                    error: "Email with id 'uuid' not found"
                    code: EMAIL_NOT_FOUND
                deleted:
                  summary: Email deleted by viewer
                  value:
                    error: "Email with id 'uuid' was deleted by viewer"
                    code: EMAIL_DELETED

    delete:
      tags:
        - Mail
      summary: Delete email
      description: |
        Soft-delete an email for a specific user. Adds the viewer's name to the `deletedBy` array.

        **Access control:** Viewer must be a participant (in `to` array OR matches `from`).

        **Idempotent:** If already deleted by this viewer, still returns 200.

        **Effects:**
        - Email no longer appears in viewer's inbox
        - Email returns 404 for viewer on direct access
        - Email still appears in thread views for all users
        - Email still appears in /investigation endpoint
      operationId: deleteEmail
      parameters:
        - name: mail_id
          in: path
          required: true
          description: UUID of the email to delete
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/viewer'
      responses:
        '200':
          description: Email deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Email deleted
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Not a participant
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "User 'viewer' is not a participant in email 'uuid'"
                code: NOT_PARTICIPANT
        '404':
          description: Email not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /investigation/{name}:
    get:
      tags:
        - Investigation
      summary: Investigate user emails
      description: |
        Get all emails involving a person, **including emails deleted by anyone**.
        For administrative/investigation purposes.

        **Pagination:** 20 items per page

        **Matching:** Returns emails where name appears in `to` array OR matches `from` field (case-insensitive)

        **Includes:** `readBy` and `deletedBy` arrays in response
      operationId: getInvestigation
      parameters:
        - name: name
          in: path
          required: true
          description: Name of person to investigate (trimmed, case-insensitive)
          schema:
            type: string
        - $ref: '#/components/parameters/page'
      responses:
        '200':
          description: Investigation results
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EmailFull'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalidName:
                  summary: Empty name
                  value:
                    error: "Name parameter cannot be empty or whitespace"
                    code: INVALID_NAME
components:
  parameters:
    viewer:
      name: viewer
      in: query
      required: true
      description: Name of the user viewing (case-insensitive, will be normalized to lowercase)
      schema:
        type: string
    page:
      name: page
      in: query
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
  schemas:
    Error:
      type: object
      required:
        - error
        - code
      properties:
        error:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
          enum:
            - EMAIL_NOT_FOUND
            - EMAIL_DELETED
            - NOT_PARTICIPANT
            - PARENT_NOT_FOUND
            - MISSING_FIELD
            - INVALID_FIELD
            - INVALID_JSON
            - INVALID_UUID
            - INVALID_PAGE
            - INVALID_NAME
            - MISSING_VIEWER
            - INVALID_VIEWER
            - UNKNOWN_PARAMETER
            - DUPLICATE_PARAMETER
            - UNKNOWN_FIELD
            - UNSUPPORTED_MEDIA_TYPE
    SendEmailRequest:
      type: object
      required:
        - to
        - from
        - subject
        - content
      properties:
        to:
          type: array
          items:
            type: string
          minItems: 1
          description: Recipient names (will be normalized to lowercase, deduplicated)
          example: ["alice", "bob"]
        from:
          type: string
          description: Sender name (will be normalized to lowercase)
          example: charlie
        subject:
          type: string
          description: Email subject line (no length limit)
          example: Meeting tomorrow
        content:
          type: string
          description: Email body, plain text only (no length limit)
          example: "Hi team, let's meet tomorrow at 10am."
        isResponseTo:
          type: string
          format: uuid
          nullable: true
          description: Parent email ID if this is a reply
    EmailSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        from:
          type: string
        to:
          type: array
          items:
            type: string
        subject:
          type: string
        timestamp:
          type: string
          format: date-time
          description: ISO 8601 UTC timestamp with Z suffix
        isResponseTo:
          type: string
          format: uuid
          nullable: true
        read:
          type: boolean
          description: True if viewer is in readBy array
    EmailDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
        from:
          type: string
        to:
          type: array
          items:
            type: string
        subject:
          type: string
        content:
          type: string
        timestamp:
          type: string
          format: date-time
        isResponseTo:
          type: string
          format: uuid
          nullable: true
        read:
          type: boolean
    ThreadSummary:
      type: object
      description: Summary of an email in a thread (no content)
      properties:
        id:
          type: string
          format: uuid
        from:
          type: string
        to:
          type: array
          items:
            type: string
        subject:
          type: string
        timestamp:
          type: string
          format: date-time
        isResponseTo:
          type: string
          format: uuid
          nullable: true
    EmailFull:
      type: object
      description: Full email object including readBy and deletedBy (for investigation)
      properties:
        id:
          type: string
          format: uuid
        from:
          type: string
        to:
          type: array
          items:
            type: string
        subject:
          type: string
        content:
          type: string
        timestamp:
          type: string
          format: date-time
        isResponseTo:
          type: string
          format: uuid
          nullable: true
        readBy:
          type: array
          items:
            type: string
          description: Names of users who have read this email
        deletedBy:
          type: array
          items:
            type: string
          description: Names of users who have deleted this email
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 10
        total_items:
          type: integer
          example: 47
        total_pages:
          type: integer
          example: 5
        has_next:
          type: boolean
        has_prev:
          type: boolean
    ThreadPagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_in_thread:
          type: integer
          example: 45
        total_pages:
          type: integer
          example: 3
        has_next:
          type: boolean
        has_prev:
          type: boolean